Add files via upload
A simple version of visual odometry. The first edition is done based on PnP solver of OpenCV, the second version is combined PnP solver and g2o optimisation. Besides, the second version is more complete with its map and continuously estimation.
A video demo of second version can be found on, which used the TMU RGBD dataset:
http://v.youku.com/v_show/id_XMzI2MTQ3MTU2NA==.html?spm=a2hzp.8244740.0.0